made pacman move continuously after pressing a direction. Will have to make him comsume the pellets

+++ Game/Character.pde
@@ -1,12 +1,12 @@
 interface Character {
   void moveX();
   void moveY();
-  void display(int x, int y);
-  int getX();
-  int getY();
-  void setX(int newX);
-  void setY(int newY);
-  int getSpeed();;
-  void setSpeed(int newSpeed);
+  void display(float x, float y);
+  float getX();
+  float getY();
+  void setX(float newX);
+  void setY(float newY);
+  float getSpeed();;
+  void setSpeed(float newSpeed);
   
 }
